Experiment_2 branch
The TODOs in the Cell class have been completed, making the game fully
functional. The settings are set to ignore the serialVersionUID warning
on the GameMain class which does not effect the program. Please refer to
paint method in the cell class and make sure all other methods and
constructors have the same level of detail when it come to
documentation. For any changes, please make a new branch.